+++ src/world/world.js
@@ -1,5 +1,5 @@
 
-import { makeNoise2D, makeNoise3D } from 'open-simplex-noise'
+import { makeNoise2D, makeNoise3D, makeNoise4D } from 'open-simplex-noise'
 
 
 var hash = require('murmur-numbers')
@@ -10,6 +10,7 @@ var cruncher = require('voxel-crunch')
 var storage = {}
 
 // plumbing
+
 
 function chunkIsStored(id) {
 	if (storage[id]) var x = true
@@ -45,11 +46,13 @@ function WorldGen() {
 	var blockIDs = {}
 	var heightNoise
 	var caveNoise
+	var caveNoise2
 	var biomeNoise1
 	var biomeNoise2
 	var biomeNoise3
 	var seed
 	var world
+	var chunksize = 24
 	this.onMessage = function (ev){
 		var msg = ev && ev.data && ev.data.msg
 		if (!msg) return
@@ -58,8 +61,10 @@ function WorldGen() {
 			blockIDs = ev.data.blocks
 			seed = ev.data.seed
 			world = ev.data.world
+			chunksize = ev.data.chunksize
 			heightNoise = makeNoise2D(seed)
 			caveNoise = makeNoise3D(seed)
+			caveNoise2 = makeNoise3D(seed*2)
 			biomeNoise1 = makeNoise2D(seed*100/17)
 			biomeNoise2 = makeNoise2D(seed*100/71)
 			biomeNoise3 = makeNoise2D(seed*100/942)
@@ -101,7 +106,7 @@ function WorldGen() {
 		if (chunkIsStored(id) && result == 'populate') populateChunk(chunk, x, y, z, id)
 		else if (chunkIsStored(id) && result == state) {}
 		else {
-			var biometable = ndarray(new Array, [24, 24])
+			var biometable = ndarray(new Array, [chunksize, chunksize])
 			generateChunk(chunk, x, y, z, id, biometable)
 			populateChunk(chunk, x, y, z, id, biometable)
 		}
@@ -116,7 +121,7 @@ function WorldGen() {
 
 	// two versions of world data
 	// `data` is an ndarray - see https://github.com/scijs/ndarray
-	var biomeSpacing = 300
+	var biomeSpacing = 200
 	async function generateChunk(chunk, x, y, z, id, biometable) {
 		var dx = chunk.shape[0]
 		var dy = chunk.shape[1]
@@ -130,7 +135,8 @@ function WorldGen() {
 				var biome = getBiome((x + i), (z + k),temperature, mountaines, biomerng)
 				biometable.set(i, k, biome) 
 				for (var j = 0; j < dy; j++) {
-					var height = getHeight(x + i, y + j, z + k, mountaines)
+					var cave = getCave(x + i, y + j, z + k)
+					var height = getHeight(x + i, y + j, z + k, cave, mountaines)
 					var block = getBlock(x + i, y + j, z + k, height, biome)
 					
 					if (block != 0) chunk.set(i, j, k, block)
@@ -142,7 +148,7 @@ function WorldGen() {
 		function getBlock(x, y, z, height, biome) {
 			if (y >= 30 && y <= height) return blockIDs.stone
 			if (y > height) return 0
-			if (y == height) {
+			if (y == height && y > 0) {
 				if (biome == 'plants' || biome == 'forest') return blockIDs.grass
 				else if (biome == 'iceland') return blockIDs.grass_snow
 				else if (biome == 'desert') return blockIDs.sand
@@ -150,18 +156,19 @@ function WorldGen() {
 			else if (y+3 < height) {
 				return blockIDs.stone
 			}
-			if (y < height) {
+			if (y < height && y > 0) {
 				if (biome == 'plants' || biome == 'forest' || biome == 'iceland') return blockIDs.dirt
 				else if (biome == 'desert') return blockIDs.sand
 			}
 		}
 
-		function getHeight(x, y, z, mountaines) {
-			var layer1 = heightNoise(x/40, z/40)*20
-			var layer2 = heightNoise(x/20, z/20)*20
-			var layer3 = heightNoise(x/10, z/10)*20
+		function getHeight(x, y, z, cave, mountaines) {
+			
+			var layer1 = heightNoise(x/40, z/40)*mountaines-cave
+			var layer2 = heightNoise(x/20, z/20)-cave
+			var layer3 = heightNoise(x/10, z/10)-cave
 
-			var noise = Math.round((layer1*20+layer2*10+layer3*5)/35) + 15
+			var noise = Math.round((layer1*20*20+layer2*20*10+layer3*20*5)/35) + 15
 			return noise
 		}
 
@@ -173,6 +180,11 @@ function WorldGen() {
 
 			else if (temperature + rnd >= 0.2 && biomerng + rnd > 0.6) return 'forest'
 			else return 'plants'
+		}
+		function getCave(x, y, z) {
+			var noise = (caveNoise(x/20, y/20, z/20) > 0.45) ? caveNoise(x/20, y/20, z/20) : 0
+			var line = Math.ceil(Math.abs(caveNoise2(x/10, y/10, z/10)))
+			return Math.round(noise*line)
 		}
 	}
 
@@ -187,15 +199,15 @@ function WorldGen() {
 				for (var j = 0; j < dy; j++) {
 					if (getBlock(i, j, k) == blockIDs.grass && hash(x+i, z+k) > 0.8 ) {
 						setBlock(i, j+1, k, blockIDs.grass_plant)
-					} else if (getBlock(i, j, k) == blockIDs.grass && hash(x+i, z+k) < 0.05 ) {
-						if (hash(x+z) >= 0.5) setBlock(i, j+1, k, blockIDs.red_flower)
+					} else if (getBlock(i, j, k) == blockIDs.grass && hash((x+i), (z+k)) < 0.05 ) {
+						if (hash(x+i+z+k) >= 0.5) setBlock(i, j+1, k, blockIDs.red_flower)
 						else setBlock(i, j+1, k, blockIDs.yellow_flower)
-					} else if (getBlock(i, j, k) == blockIDs.grass && hash(x+i*seed, z+k*seed) < 0.01 && biome == 'plants' ) {
+					} else if (getBlock(i, j, k) == blockIDs.grass && hash((x+i)*seed, (z+k)*seed) < 0.002 && biome == 'plants' ) {
 						tree(i, j, k, y+j)
-					} else if (getBlock(i, j, k) == blockIDs.grass && hash(x+i*seed, z+k*seed) < 0.8 && biome == 'forest' ) {
+					} else if (getBlock(i, j, k) == blockIDs.grass && hash((x+i)*seed, (z+k)*seed) < 0.02 && biome == 'forest' ) {
 						tree(i, j, k, y+j)
-					} else if (getBlock(i, j, k) == blockIDs.sand && getBlock(i, j+1, k) == 0  && hash(x+i*seed, z+k*seed) < 0.01 && biome == 'desert' ) {
-						var w = 3 (hash(i, k) >= 0.5) ? 1 : 0
+					} else if (getBlock(i, j, k) == blockIDs.sand && getBlock(i, j+1, k) == 0  && hash((x+i)*seed, (z+k)*seed) < 0.01 && biome == 'desert' ) {
+						var w = 3 + (hash(i, k) >= 0.5) ? 1 : 0
 						for (var h = 0; h <= w; h++) {
 							setBlock(i, j+h+1, k, blockIDs.cactus)
 						}
@@ -220,7 +232,25 @@ function WorldGen() {
 
 
 		function getBlock(x, y, z) {
-			return chunk.get(x, y, z)
+			if (x >= 0 && x < dx && y >= 0 && y < dy && z >= 0 && z < dz) { return chunk.get(x, y, z) }
+			else { 
+				var splitID = id.split("|")
+				var tempX = (x >= 0 && x < dx ) ? x : ((x < 0) ? x + dx : x - dx )
+				var tempY = (y >= 0 && y < dy ) ? y : ((y < 0) ? y + dy : y - dy )
+				var tempZ = (z >= 0 && z < dz ) ? z : ((z < 0) ? z + dz : z - dz )
+
+				for (var w = 0; w < 3; w++) {splitID[w] = parseInt(splitID[w])}
+
+				splitID[0] = (x >= 0 && x < dx ) ? splitID[0] : ((x < 0) ? splitID[0]-1 : splitID[0]+1)
+				splitID[1] = (y >= 0 && y < dy ) ? splitID[1] :((y < 0) ? splitID[1]-1 : splitID[1]+1)
+				splitID[2] = (z >= 0 && z < dz ) ? splitID[2] : ((z < 0) ? splitID[2]-1 : splitID[2]+1)
+
+				var tempID = splitID[0] + '|' + splitID[1] + '|' + splitID[2] + '|' + splitID[3]
+				
+				var tempChunk = new ndarray( new Uint16Array(chunksize * chunksize * chunksize), [chunksize, chunksize, chunksize] )
+				if (chunkIsStored(tempID)) { var state = retrieveChunk(tempID, tempChunk) }
+				return tempChunk.get(tempX, tempY, tempZ, block)
+			}
 		}
 
 		async function setBlock(x, y, z, block) {
@@ -243,7 +273,7 @@ function WorldGen() {
 				//sendDebug(tempID)
 				//sendDebug([tempX, tempY, tempZ])
 				
-				var tempChunk = new ndarray( new Uint16Array(24 * 24 * 24), [24, 24, 24] )
+				var tempChunk = new ndarray( new Uint16Array(chunksize * chunksize * chunksize), [chunksize, chunksize, chunksize] )
 				if (chunkIsStored(tempID)) { var state = retrieveChunk(tempID, tempChunk) }
 				tempChunk.set(tempX, tempY, tempZ, block)
 				saveChunk(tempID, tempChunk)